+++ das_client/app/lib/pages/journey/journey_screen/widgets/departure_process_dialog.dart
@@ -170,6 +170,4 @@ class DepartureProcessDialog extends StatelessWidget {
       ),
     );
   }
-
-  void _doNothing() {}
 }