+++ das_client/app/test/pages/journey/journey_screen/view_model/checklist_departure_process_view_model_test.dart
@@ -368,7 +368,7 @@ void main() {
       });
       processStreams(fakeAsync: testAsync);
 
-      expect(modelRegister.where((m) => m is ChecklistDepartureProcessDisabled), isEmpty);
+      expect(modelRegister.whereType<ChecklistDepartureProcessDisabled>(), isEmpty);
     });
   });
 }