+++ das_client/sfera/lib/src/data/local/drift_sfera_local_database_service.dart
@@ -1,8 +1,12 @@
 // coverage:ignore-file
 
+import 'dart:io';
+
 import 'package:drift/drift.dart';
-import 'package:drift_flutter/drift_flutter.dart';
+import 'package:drift/native.dart';
 import 'package:logging/logging.dart';
+import 'package:path/path.dart' as p;
+import 'package:path_provider/path_provider.dart';
 import 'package:sfera/src/data/dto/journey_profile_dto.dart';
 import 'package:sfera/src/data/dto/segment_profile_dto.dart';
 import 'package:sfera/src/data/dto/train_characteristics_dto.dart';
@@ -12,9 +16,9 @@ import 'package:sfera/src/data/local/tables/segment_profile_table.dart';
 import 'package:sfera/src/data/local/tables/train_characteristics_table.dart';
 import 'package:sfera/src/model/sfera_db_metrics.dart';
 
-part 'drift_local_database_service.g.dart';
+part 'drift_sfera_local_database_service.g.dart';
 
-final _log = Logger('SferaDriftLocalDatabaseService');
+final _log = Logger('DriftSferaLocalDatabaseService');
 
 @DriftDatabase(
   tables: [
@@ -23,19 +27,32 @@ final _log = Logger('SferaDriftLocalDatabaseService');
     TrainCharacteristicsTable,
   ],
 )
-class DriftLocalDatabaseService extends _$DriftLocalDatabaseService implements SferaLocalDatabaseService {
+class DriftSferaLocalDatabaseService extends _$DriftSferaLocalDatabaseService implements SferaLocalDatabaseService {
   static const int cleanupDays = 2;
-  static DriftLocalDatabaseService? _instance;
+  static DriftSferaLocalDatabaseService? _instance;
 
-  static DriftLocalDatabaseService get instance {
-    _instance ??= DriftLocalDatabaseService._();
+  static DriftSferaLocalDatabaseService get instance {
+    _instance ??= DriftSferaLocalDatabaseService._();
     return _instance!;
   }
 
-  DriftLocalDatabaseService._() : super(driftDatabase(name: 'sfera_db'));
+  DriftSferaLocalDatabaseService._() : super(_openConnection());
+
+  static QueryExecutor _openConnection() => LazyDatabase(() async {
+    final dbFolder = await getApplicationDocumentsDirectory();
+    final file = File(p.join(dbFolder.path, 'sfera_db.sqlite'));
+    return NativeDatabase.createInBackground(
+      file,
+      setup: (db) {
+        // settings to improve writing and reduce blocking actions
+        db.execute('PRAGMA journal_mode = WAL;');
+        db.execute('PRAGMA synchronous = NORMAL;');
+      },
+    );
+  });
 
   @override
-  int get schemaVersion => 2;
+  int get schemaVersion => 3;
 
   @override
   MigrationStrategy get migration => MigrationStrategy(
@@ -47,6 +64,11 @@ class DriftLocalDatabaseService extends _$DriftLocalDatabaseService implements S
         await m.drop(trainCharacteristicsTable);
         await m.create(trainCharacteristicsTable);
       }
+
+      if (from < 3) {
+        await m.drop(journeyProfileTable);
+        await m.create(journeyProfileTable);
+      }
     },
   );
 
@@ -55,9 +77,14 @@ class DriftLocalDatabaseService extends _$DriftLocalDatabaseService implements S
     String company,
     String operationalTrainNumber,
     DateTime startDate,
-  ) async => _jpManager
-      .filter((f) => f.company(company) & f.operationalTrainNumber(operationalTrainNumber) & f.startDate(startDate))
-      .getSingleOrNull();
+  ) async {
+    final journeyProfiles = await _jpManager
+        .filter((f) => f.company(company) & f.operationalTrainNumber(operationalTrainNumber) & f.startDate(startDate))
+        .orderBy((o) => o.version.desc())
+        .get();
+
+    return journeyProfiles.firstOrNull;
+  }
 
   @override
   Future<SegmentProfileTableData?> findSegmentProfile(String spId, String majorVersion, String minorVersion) async =>
@@ -85,29 +112,41 @@ class DriftLocalDatabaseService extends _$DriftLocalDatabaseService implements S
 
   @override
   Future<void> saveJourneyProfile(JourneyProfileDto journeyProfile) async {
-    final existingProfile = await findJourneyProfile(
-      journeyProfile.trainIdentification.otnId.company,
-      journeyProfile.trainIdentification.otnId.operationalTrainNumber,
-      journeyProfile.trainIdentification.otnId.startDate,
-    );
-
-    final journeyProfileCompanion = journeyProfile.toCompanion(id: existingProfile?.id);
-
+    final otnId = journeyProfile.trainIdentification.otnId;
     _log.fine(
-      'Writing journey profile to db company=${journeyProfileCompanion.company} operationalTrainNumber=${journeyProfileCompanion.operationalTrainNumber} startDate=${journeyProfileCompanion.startDate}',
+      'Writing journey profile to db company=${otnId.company} operationalTrainNumber=${otnId.operationalTrainNumber} startDate=${otnId.startDate}',
     );
-    await journeyProfileTable.insertOnConflictUpdate(journeyProfileCompanion);
+    await journeyProfileTable.insertOnConflictUpdate(journeyProfile.toCompanion());
   }
 
   @override
-  Future<void> saveSegmentProfile(SegmentProfileDto segmentProfile) async {
-    await segmentProfileTable.insertOnConflictUpdate(segmentProfile.toCompanion());
+  Future<void> saveBulkJourneyProfiles(Iterable<JourneyProfileDto> journeyProfiles) async {
+    Insertable<JourneyProfileTableData> mapToCompanion(JourneyProfileDto jp) {
+      final otnId = jp.trainIdentification.otnId;
+      _log.fine(
+        'Writing journey profile to db company=${otnId.company} operationalTrainNumber=${otnId.operationalTrainNumber} startDate=${otnId.startDate}',
+      );
+      return jp.toCompanion();
+    }
+
+    _jpManager.bulkCreate((_) => journeyProfiles.map(mapToCompanion), mode: .insertOrReplace);
   }
 
   @override
-  Future<void> saveTrainCharacteristics(TrainCharacteristicsDto trainCharacteristics) async {
-    await trainCharacteristicsTable.insertOnConflictUpdate(trainCharacteristics.toCompanion());
-  }
+  Future<void> saveSegmentProfile(SegmentProfileDto segmentProfile) =>
+      segmentProfileTable.insertOnConflictUpdate(segmentProfile.toCompanion());
+
+  @override
+  Future<void> saveBulkSegmentProfiles(Iterable<SegmentProfileDto> segmentProfiles) =>
+      _spManager.bulkCreate((_) => segmentProfiles.map((sp) => sp.toCompanion()), mode: .insertOrReplace);
+
+  @override
+  Future<void> saveTrainCharacteristics(TrainCharacteristicsDto trainCharacteristics) =>
+      trainCharacteristicsTable.insertOnConflictUpdate(trainCharacteristics.toCompanion());
+
+  @override
+  Future<void> saveBulkTrainCharacteristics(Iterable<TrainCharacteristicsDto> trainCharacteristics) =>
+      _tcManager.bulkCreate((_) => trainCharacteristics.map((tc) => tc.toCompanion()), mode: .insertOrReplace);
 
   $$TrainCharacteristicsTableTableTableManager get _tcManager => managers.trainCharacteristicsTable;